Adding updates to vehicle physics

+++ Assets/WestCoastPhysics/DevTools/Observer/Observer.cs
@@ -17,16 +17,16 @@ public class EventArguments
 {
     public string eventMessage{ get; set; }
     public string extendedMessage{ get; set; }
-    public int extendedMessageNumber{ get; private set; }
+    public float extendedMessageNumber{ get; private set; }
     public Vector3 objectCoordinates{ get; private set; }
 
     public EventArguments( string newEventMessage, string newExtendedMessage, Vector3 coordinates )
 	{
-        int newInteger;
+        float newInteger;
         eventMessage = newEventMessage;
         extendedMessage = newExtendedMessage;
         objectCoordinates = coordinates;
-        if( int.TryParse( newExtendedMessage, out newInteger ) )
+        if( float.TryParse( newExtendedMessage, out newInteger ) )
 		{
             extendedMessageNumber = newInteger;
         }
